+++ docs/source/changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.4.3 (202Y-MM-DD)
 - drop support for python 3.10
+- add support for hashing pandas dataframes and arrow tables
+- use pyarrow for hashing of polars dataframes, fixing issues with categorical columns
 
 ## 0.4.2 (2025-10-10)
 - support python 3.14